Useless double.

+++ src/auxiliary.cpp
@@ -303,8 +303,8 @@ Eigen::VectorXf zscore(const Eigen::VectorXf &input) {
     if (input.size() == 0)
         return Eigen::VectorXf();
 
-    double mean = arithmetic_mean(input);
-    double stdev = standard_deviation(input);
+    float mean = arithmetic_mean(input);
+    float stdev = standard_deviation(input);
 
     Eigen::VectorXf transformed(input.size());
     for (int i = 0; i < input.size(); i++) {
